requirements.txt
Flask==2.3.3
ultralytics==8.0.196
opencv-python==4.8.1.78
pandas==2.1.0
openpyxl==3.1.2
Pillow==10.0.0
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
Werkzeug==2.3.7
PyYAML==6.0.1
